causallm_subfunction testing
